+++ changelog.md
@@ -1,6 +1,10 @@
 # Changelog
 📝 following beta format X.Y.Z where Y = breaking change and Z = feature and fix. Later => FAIL.FEATURE.FIX
 
+## 0.6.1(2023-11-07)
+- Feature: Added returnNulls to explicitly return nulls of queried fields when empty
+- Types: Added query config type
+
 ## 0.6.0(2023-11-02)
 - Chore: Updated to last version of typeDB, might not work with older versions